Complite Data Task 1

+++ DATA_Task1_parseBook/Model/Book.cs
@@ -2,7 +2,7 @@
 
 public class Book
 {
-    public ulong Id { get; set; }
+    public decimal Id { get; set; }
     public string Title { get; set; }
     public string Author { get; set; }
     public string Genre { get; set; }
